+++ src/payment-form/2.js
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export default () => (
+  <h1>page 2</h1>
+)
