+++ src/payment-form/index.js
@@ -1,76 +1,25 @@
-import React, { useState } from 'react';
-import Validator from '../Validator';
-import H1 from '../components/H1';
-import Input from '../components/Input';
-import Field from '../components/Field';
-import Label from '../components/Lable';
-import FieldLabel from '../components/Field-Label';
-import FieldHint from '../components/Field-Hint';
-import FieldError from '../components/Field-Error';
-import ErrorSummary from '../components/Error-Summary';
-import Submit from '../components/Submit';
+import React from 'react';
+import {
+  BrowserRouter as Router,
+  Route,
+  Redirect
+} from "react-router-dom";
 
-let originalTitle = document.title;
+import Page1 from './1';
+import Page2 from './2';
+
+const components = [Page1, Page2];
 
 export default () => {
-  const validator = new Validator();
-  const [email, setEmail] = useState('');
-  const [errors, setErrors] = useState(new Validator().errors);
-
-  validator.add(
-    {
-      key: 'email',
-      valid() {
-        return email.length > 0
-      },
-      errorMessage: 'メールアドレスを入力してください。',
-    }
-  );
-  validator.add(
-    {
-      key: 'email',
-      valid() {
-        return email.includes('@')
-      },
-      errorMessage: '@を入れてください。',
-    },
-  );
-
-  function handleSubmit(e) {
-    e.preventDefault();
-    setErrors(new Validator().errors);
-
-    if (validator.invalid()) {
-      const errors = validator.errorMessages();
-      setErrors(errors);
-      document.title = `(${errors.count()}件のエラー) - ${originalTitle}`;
-    }
-  };
-
-  let emailErrorMessage = errors.getMessage('email');
-
   return (
-    <>
-      <ErrorSummary errors={errors} />
-      <H1>支払いフォーム</H1>
-      <form noValidate onSubmit={(e) => handleSubmit(e)}>
-        <Field>
-          <Label htmlFor="email">
-            <FieldLabel>メールアドレス</FieldLabel>
-            <FieldHint>こちらにご注文の確認メールをお送りします。</FieldHint>
-            <FieldError message={emailErrorMessage} />
-          </Label>
-          <Input
-            type="email"
-            id="email"
-            name="email"
-            value={email}
-            aria-invalid={false}
-            onChange={(e) => setEmail(e.target.value)}
-          />
-        </Field>
-        <Submit type="submit" value="続ける" />
-      </form>
-    </>
+    <Router>
+      <Route exact path="/payment-form">
+        <Redirect to="/payment-form/1" />
+      </Route>
+
+      {components.map((component, i) => (
+        <Route key={i} path={`/payment-form/${i+1}`} component={component} />
+      ))}
+    </Router>
   )
 }
